+++ src/sdc11073/consumer/consumerimpl.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import copy
 import functools
 import logging
+import socket
 import ssl
 import time
 import traceback
@@ -199,6 +200,7 @@ class SdcConsumer:
                  request_chunk_size: int = 0,
                  socket_timeout: int = 5,
                  force_ssl_connect: bool = False,
+                 use_full_qualified_name: bool = False
                  ):
         """Construct a SdcConsumer.
 
@@ -216,6 +218,7 @@ class SdcConsumer:
                                   False: if ssl_context_container is provided, consumer first tries an
                                          encrypted connection, and if this raises an SSLError,
                                          it tries an unencrypted connection
+        :param use_full_qualified_name: if true use the full_qualified_name for subscriptions, default (false) is to use numerical IP
         """
         if not device_location.startswith('http'):
             raise ValueError('Invalid device_location, it must be match http(s)://<netloc> syntax')
@@ -294,6 +297,7 @@ class SdcConsumer:
 
         self._msg_converter = MessageConverterMiddleware(
             self.msg_reader, self.msg_factory, self._logger, self._services_dispatcher)
+        self._use_full_qualified_name = use_full_qualified_name
 
     def set_mdib(self, mdib: ConsumerMdib | None):
         """SdcConsumer sometimes must know the mdib data (e.g. Set service, activate method)."""
@@ -340,7 +344,10 @@ class SdcConsumer:
         if self._http_server is None:
             return ''
         p = urlparse(self._http_server.base_url)
-        tmp = f'{p.scheme}://{self._network_adapter.ip}:{p.port}{p.path}'
+        if self._use_full_qualified_name:
+            tmp = f'{p.scheme}://{socket.getfqdn()}:{p.port}{p.path}'
+        else:
+            tmp = f'{p.scheme}://{self._network_adapter.ip}:{p.port}{p.path}'
         sep = '' if tmp.endswith('/') else '/'
         tmp = f'{tmp}{sep}{self.path_prefix}/'
         return tmp